Address review feedback on onboarding page
1. Deduplicate setup queries: useAutopilotPageData now delegates to
   useSetupStatus instead of duplicating the same 5 queries.

2. Fix redirect hop for returning users: middleware points back to
   /autopilot so returning users with a session cookie skip the
   onboarding API round-trip. Only fresh login and invite-accept
   flows route through /onboarding.

3. Move SetupChecklist to shared location: extracted from
   components/autopilot/ to components/setup-checklist.tsx since
   it's now used by both autopilot and onboarding. Old path
   re-exports for backwards compatibility.

4. Match loading state text pattern with rest of the app.
