+++ frontend/src/middleware.ts
@@ -4,12 +4,11 @@ import type { NextRequest } from "next/server";
 export function middleware(request: NextRequest) {
   const sessionToken = request.cookies.get("session_token");
 
-  // If user has a session cookie and is visiting /login, redirect to /onboarding.
-  // The onboarding page checks setup status and redirects to /autopilot once complete.
+  // If user has a session cookie and is visiting /login, redirect to /autopilot.
   // The cookie is not validated here — if it's expired or invalid, the dashboard's
   // auth check will catch it and redirect back to /login.
   if (sessionToken?.value) {
-    return NextResponse.redirect(new URL("/onboarding", request.url));
+    return NextResponse.redirect(new URL("/autopilot", request.url));
   }
 
   return NextResponse.next();
